@@ -1,14 +1,14 @@
 module github.com/gitlabhq/terraform-provider-gitlab
 
-go 1.17
+go 1.19
 
 require (
 	github.com/hashicorp/go-cty v1.4.1-0.20200414143053-d3edf31b6320
 	github.com/hashicorp/go-retryablehttp v0.7.1
-	github.com/hashicorp/terraform-plugin-sdk/v2 v2.17.0
+	github.com/hashicorp/terraform-plugin-sdk/v2 v2.19.0
 	github.com/mitchellh/hashstructure v1.1.0
-	github.com/onsi/gomega v1.19.0
-	github.com/xanzy/go-gitlab v0.68.2
+	github.com/onsi/gomega v1.20.1
+	github.com/xanzy/go-gitlab v0.73.1
 )
 
 require (
