@@ -19,6 +19,11 @@ func gitlabRepositoryFileGetSchema() map[string]*schema.Schema {
 			Required:    true,
 			ForceNew:    true,
 		},
+		"content": {
+			Description: "File content. If the content is not yet base64 encoded, it will be encoded automatically. No other encoding is currently supported, because of a [GitLab API bug](https://gitlab.com/gitlab-org/gitlab/-/issues/342430).",
+			Type:        schema.TypeString,
+			Required:    true,
+		},
 		"ref": {
 			Description: "The name of branch, tag or commit.",
 			Type:        schema.TypeString,
@@ -38,11 +43,6 @@ func gitlabRepositoryFileGetSchema() map[string]*schema.Schema {
 			Description: "The file content encoding.",
 			Type:        schema.TypeString,
 			Computed:    true,
-		},
-		"content": {
-			Description: "File content. If the content is not yet base64 encoded, it will be encoded automatically. No other encoding is currently supported, because of a [GitLab API bug](https://gitlab.com/gitlab-org/gitlab/-/issues/342430).",
-			Type:        schema.TypeString,
-			Required:    true,
 		},
 		"content_sha256": {
 			Description: "File content sha256 digest.",
